cloud_function/requirements.txt
# Function dependencies, for example:
pytest >= 7.1.2
python-dotenv >= 0.20.0
requests-cache >= 0.9.4
click >= 8.1.2
google-cloud-bigquery >= 3.1.0
git+https://github.com/urquha/py_forecast@master